crawling/requirements.txt
beautifulsoup4==4.12.2
greenlet==3.0.1
mysql-connector-python==8.2.0
numpy==1.26.1
pandas==2.1.2
protobuf==4.21.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
